ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ML - Face Detection + Embedding
deepface==0.0.93
retinaface==0.0.17
tf-keras>=2.16.0
onnxruntime==1.20.1
torch>=2.1.0
torchvision>=0.16.0

# Image processing
opencv-python-headless==4.10.0.84
numpy>=1.24.0
Pillow>=10.0.0

# Schema validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest==8.3.4
httpx==0.28.1
